Require Jenkins 2.452.4 or newer
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends Jenkins 2.452.4 as a good baseline for Jenkins versions.
It allows us to use Java 11 or Java 17 and includes a fix for critical
security vulnerabilities.

https://stats.jenkins.io/pluginversions/folder-properties.html shows
that 33% of installations of the most recent release are already running
Jenkins 2.452.4.

Uses the plugin bill of materials to simplify dependency management
https://www.jenkins.io/doc/developer/tutorial-improve/use-plugin-bill-of-materials/

Moves the plugin description to the jelly file where it will be read by
the Jenkins update center.

Removes unnecessary dependencies from pom file.

Adapts one test to a change in exception handling.