Remove incorrect withFolderProperties from options block
Declarative Pipeline directive generator incorrectly offers a
withFolderProperties setting for the option block.  When that setting is
used, the declarative Pipeline fails with an error that the setting is
not recognized.

Use SynchronousNonBlockingStepExecution instead of deprecated
AbstractSynchronousNonBlockingStepExecution.  The deprecated
method uses Guice and we want to reduce the use of Guice in Jenkins.