Fix JSR-305 annotations
Annotations for Nonnull, CheckForNull, and several others were proposed
for Java as part of dormant Java specification request JSR-305. The
proposal never became a part of standard Java. Jenkins plugins should
switch from using JSR-305 annotations to use Spotbugs annotations that
provide the same semantics.

Refer to the mailing list discussion from James Nord at
https://groups.google.com/g/jenkinsci-dev/c/uE1wwtVi1W0/m/gLxdEJmlBQAJ .

https://docs.openrewrite.org/recipes/jenkins/javaxannotationstospotbugs
is the OpenRewrite recipe that performs the transformation.

https://www.jenkins.io/doc/developer/tutorial-improve/replace-jsr-305-annotations/
is the improve a plugin tutorial that recommends the transformation and
describes the rationale.